% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_links.R
\name{assignment_links}
\alias{assignment_links}
\title{Get links for page assignment}
\usage{
assignment_links(url, cookies)
}
\arguments{
\item{url}{\itemize{
\item url of the assignment on gradescope
}}

\item{cookies}{\itemize{
\item list object containing login cookies for gradescope, obtained using \code{get_login_cookies}.
}}
}
\description{
Get links for page assignment
}
