% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_page_assignments.R
\name{get_page_assignments}
\alias{get_page_assignments}
\title{Get assigned pages for each question and subquestion for an assignment}
\usage{
get_page_assignments(link)
}
\arguments{
\item{link}{\itemize{
\item url of the question assignment page, e.g. those returned by assignment_links
}}
}
\description{
Get assigned pages for each question and subquestion for an assignment
}
