% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_assignment.R
\name{upload_assignment}
\alias{upload_assignment}
\title{Upload assignment}
\usage{
upload_assignment(url, student, file, cookies)
}
\arguments{
\item{url}{\itemize{
\item url of the assignment on gradescope
}}

\item{student}{\itemize{
\item student's name (exactly) as it appears in the gradescope course roster
}}

\item{file}{\itemize{
\item path of the local file to upload
}}

\item{cookies}{\itemize{
\item list object containing login cookies for gradescope, obtained using \code{get_login_cookies}.
}}
}
\description{
Upload assignment
}
